Revert move of Run() to google::cloud::functions, too early
